Data fix: Origin dates
The results of running fix-dates.xsl on all Fihrist files were 6720 out of 11842 TEI files modified. No changes to the text of the origDate elements, only either modifying/adding normalized and machine-readable attributes, or removing the origDate tags altogether when they contain a statement that there is no date. Closes #17.