Deleting one works authority entry
This entry was created because an msItem in Kings_46.xml had a blank title. That title has since been filled in, so a new entry has been created for it, making this one superfluous. It never appeared on the web site so there is no need to move it to works_deletions.xml.