Data fix: Languages
The mainLang and otherLangs attributes must be ISO 639 codes (parts 1, 2 or 3) and the text of the textLang element should contain a description for human readers. This fixes the 180 instances in all current Fihrist files where one or both of those stipulations were not true. It does not change one kind of ISO 639 code to another.