Data cleanup project - improved versions of persons authority files
persons_base.xml and persons_deletions.xml are being regenerated in a one-off process, from a combination of two sources: the id remapping database used for the data cleanup project and the post-data-cleanup versions of the manuscript TEI files.

I have fixed some errors in the way variant names and manuscript references (comments) were being constructed in this one-off process.