Data cleanup project - corrected works_base authority file
works_base.xml is regenerated in a one-off process, from a combination of two sources: the id remapping database used for the data cleanup project and the post-data-cleanup versions of the manuscript TEI files.

The file has been fixed to make it valid TEI: using author or editor rather than person elements and using key instead of xml:id.

I have fixed some errors in the way variant titles and manuscript references (comments) were being constructed in this one-off process.

Minor changes were made to Oxford Bod. Elliott 135 and 167 – forcing a key value for an initial title element not to be blank when a later sibling provided the work id.