Data cleanup project - stripped out role 'aut' from author elements
Stripped out logically (almost) redundant 'aut' role code from every author element.

Some author elements will still have other roles assigned, for example in Eton 282 (manuscript_15421) there is an author assigned the role 'scr' (scribe) - this is a holograph manuscript.

I say these 'aut' codes were logically ALMOST redundant for the reason that just because a person appears in an author element within a msItem this does NOT mean they are necessarily author of the main work of the msItem. For example, author elements are sometimes used as children of note/listBibl/bibl elements, in a note that refers to a related work. See, for example, Cam Uni Add 3590 (manuscript_9265) item 6 where Juwaynī is in an author element tagged with role 'ant' - here his role relative to the main work of the msItem is NOT as author but as bibliographic antecedent .