Data cleanup project - divide role code [dpc] between [dpc] and [subject]
Previously the role code [dpc] was used for two different sorts of 'depicted' person:

(1) persons depicted in a work of art (illustration)
(2) persons who are a subject of a text

The second subgroup have now been reassigned to a new non-standard (non-MARC) code created for this purpose: [subject]

Persons now tagged with role code [subject] are related to the logical work (regardless of depictions in any particular manuscript copy) and will be listed under the catch-all non-authorial 'contributors' to the work.

Persons with role [dpc] are only related to a specific manuscript (even if in principle they are likely to also be 'subject' of the logical work they are not usually marked up as such).