Data cleanup project - fixed up some cataloguing/markup errors in manuscript TEI files
Fixed various cataloguing/mark-up errors in manuscript TEI files - errors that were spotted during work on the data cleanup project:

The main type of error was cataloguers using the wrong person id for the correct person. (Such cases could not be fixed using id remapping at the global level since the error only occurred in one or a few places.) In a few cases (e.g. BL IO Islamic 4711, King's 250) errors were found and corrected in the name of the person as well as the person_id.

The second main type of problem here fixed is where different title elements for a given work (msItem) had been tagged with different work ids. In such cases, the key value for subsequent titles should now match the key value of the first title element.

In a few cases (e.g. Cambridge Add 327, Oxford A Walker 92) errors were found and corrected in the title of the work.

Elsewhere a title had been tagged using a persName element instead of a title element (Oxford, Bodl Or 434).

Note that, unlike id remapping and role assignment, these manual fixes have NOT been tagged with a change attribute - the only evidence of these changes is in the Github comparisons.