Data cleanup project - tidying up persons authority file
A vast number of minor modifications to 'display' names in the persons authority file - a step towards normalizing the display format for persons. (Display names are independent of the names used in ms TEI files, even if those provided the initial form).

A few manuscript TEI files have been modified, mostly due to changes in person ids, where further cases of duplication have been spotted or where there was an error in the assigned id.