Data cleanup project - tidying up works authority file
Modifications to numerous uniform (display) titles for work (bibl) entries in the works authority file.

Mostly a matter of removing transcriptions wherever possible, either by deleting the transcription (where it was mere duplication) or by replacing it with a transliteration. In many cases, a transliteration pre-existed in the authority entry as a 'variant' title in which case the 'uniform' and 'variant' titles were simply switched.

This is a small step towards normalizing the display format for works. Uniform (display) titles are independent of the titles used in ms TEI files, even if those provided the initial form. Assuming the original uniform title still exists in some ms TEI file, that will remain (with unmodified text) in the authority file, but it will now be tagged as a 'variant' title.