Data cleanup project - fixed dead catalogue links
This batch of changes involves replacing thousands of dead catalogue links by links to specific pages in current digital editions of the relevant catalogues at Cambridge Uni or SOAS. E.g.
<ref target="http://www.lib.cam.ac.uk/arabic_catalogues/browne1922/view.php?id=54">no. 501</ref>
(where the target is a dead link) is modified to
<ref target="https://cudl.lib.cam.ac.uk/view/PR-A-00233-00034-00002/51">no. 501</ref>
(where the target should point to the correct page and so the hyperlink on Fihrist should now function correctly)

List of the catalogues dealt with:
Cam Uni:
1.	Browne 1896 – now https://cudl.lib.cam.ac.uk/view/PR-A-00233-00040
2.	Browne 1900 – now https://cudl.lib.cam.ac.uk/view/PR-A-00233-00034-00001
3.	Browne 1922 – now https://cudl.lib.cam.ac.uk/view/PR-A-00233-00034-00002
4.	Arberry 1952 – now https://cudl.lib.cam.ac.uk/view/PR-A-00233-00034-00003
5.	Nicholson 1932 – now https://cudl.lib.cam.ac.uk/view/PR-A-00233-00004

SOAS:
6.	Gacek 1981 – now https://digital.soas.ac.uk/LOAA005309/00001
7.	Blumhardt / Mackenzie 1965 – now https://digital.soas.ac.uk/LOAC000067/00001

While doing this, a number of links to Blumhardt / Mackenzie 1965 which were not strictly dead but which pointed only to the catalogue cover page have also been modified so as to point to the specific page. E.g.
<ref target="http://digital.soas.ac.uk/l/LOAC000067/">Mackenzie no.40</ref>
is modified to
<ref target="https://digital.soas.ac.uk/LOAC000067/00001/50x">Mackenzie no.40</ref>