Reversing unwanted modifications to whitespace and empty elements – and fixing data corruption related to Arabic/Persian letter mīm
This batch of changes is mostly concerned with reversing some unnecessary whitespace modifications that resulted from actioning markup requests (i.e. in the previous commit). We also reverse the introduction of some unnecessary abbreviated forms for empty elements (whereby <locus…></locus> was being replaced with <locus…./> for example).

The net effect of this commit and the previous commit should mean that very few whitespace and empty element changes now appear in this branch (relative to the master). And the branch should now do only do what it was intended to do: to wrap thousands of free format provenance-related names in persName elements (with attributes) and to mark the changes as #prov-markup (linked to an explanatory <change> element under <revisionDesc>) while changing as little as possible of the whitespace.

Inevitably some pieces of formatted XML will no longer look quite so well-aligned in the manuscript descriptions when viewed in an editor – but such formatting is semantically neutral (in XML) and it should have no effect on the HTML displayed on the Fihrist site (or on anything else).

There were a few cases of data corruption in the previous commit. Specifically, ./the university of Manchester/Persian_MS_301.xml (and a few others mss) had problems with mīm in some of the Persian or Arabic words – as a result of which the mīm would have been wrongly displayed in the HTML, disconnected from the previous or following letter to which it should be joined. The obscure bug that caused this has now been identified and firmly swatted, so any related data corruption should be reversed in the current commit. Other examples of the mīm problem (now fixed) include: Wellcome MS Arabic 430 and 442.

There were also two cases where a strange initial control character in the ms XML file was causing problems for our whitespace modification reversal routines. These were WadhMinasian_573.xml and WadhMinasian_437.xml. We have removed the offending character in these two cases (the character is shown as a red dot in Github desktop).