All scripted modifications reapplied following reversion to master on 2021-11-06
All markup requests now reapplied - and authority files regenerated - having previously 'reverted' this branch to match the current master (as of 2021-11-06).

When viewed in Github Desktop, the diff between this and the previous commit on this branch should display the net affect of all scripted changes relative to the current master.