Initial commit to branch qq-cam-uni-markup
These changes to Cam Uni Qq (Burkhardt) collection fill in data fro extent (total folios) and layout (dimensions).

The data was suppled by Yasmin/Jeremy in Excel. I, Julian, use script to extract the data and make the modifications to the TEI XML.

At this point the data is not complete so there will have to be further commits in the coming days and before we can attempt to merge this branch with master...