Publish 849 new British Library Persian records (along with 18 replacements)
British Library Persian manuscripts.

This batch includes 867 manuscript files, of which 849 are entirely new to FIHRIST.

There are also18 replacements for manuscripts that are already on FIHRIST:

•	Uk_Delhi_Persian_982
•	Uk_Delhi_Persian_1145
•	Uk_Delhi_Persian_1169
•	Uk_Delhi_Persian_1183
•	Uk_IO_Islamic_3677
•	Uk_IO_Islamic_3913
•	Uk_IO_Islamic_3958
•	Uk_IO_Islamic_4063
•	Uk_IO_Islamic_4401
•	Uk_IO_Islamic_4511
•	Uk_IO_Islamic_4634
•	Uk_Or_15497
•	Uk_Or_15597
•	Uk_Or_15657
•	Uk_Or_15658
•	Uk_Or_15661
•	Uk_Or_15662
•	Uk_Or_15759

The manuscript files are associated with ids: manuscript_16713 to manuscript_17580, except in case of the 18 replacements (where the pre-existing manuscript number is reused).

Includes new version of 4 authority files:
•	persons_additions.xml
•	persons_base.xml
•	works_additions.xml
•	works_base.xml

Each includes manuscript-msItem-specific comments indicating where a person or work is found. The persons_additions file contains a block of persons new to FIHRIST - where these are found in VIAF the VIAF/LC/ISNI numbers are included.

The new FIHRIST person ids are in range person_f7001 to person_f7791; approximately 785 total new persons; there are a further 155 new persons that were found in VIAF but not previously in FIHRIST.

The new work ids are in the range work_21601 to work_22261 (approx. 660 new works).

Note that no subject authority files are included - these should be regenerated when FIHRIST is next re-indexed.