BL: tidying up Persian records previously published to FIHRIST
The 'legacy' BL files that have been on FIHRIST for years are here tidied up.

Each has been passed through XML::Tidy (using Perl) to tidy up XML formatting in what should be semantically neutral ways. (E.g. a space is inserted between a final attribute and a closing slash in an element without children; any final empty line is removed from the file, etc).

Numerous changes were made to ensure that all files validate in Oxygen against two schemas: msdesc_4_2_2.rng and enrich_BL.schematron

Inserted a newly standardized availability statement, with a link to Finding and ordering... The target page (https://www.bl.uk/help/find-persian-manuscripts) contains useful bibliographical information (lists of BL catalogues with links to those that have been digitized) as well as information on ordering manuscripts.  Since this info is in the availability element a user would find this link if they were to click on the new "How to request this item" button on the FIHRIST manuscript pages.

Note that the new  "View Online" button (and a hyperlink under Digital Images) should also display on FIHRIST manuscript pages for the 38 legacy files for which we have inserted links to digital surrogates. Clicking on either of these should take one to a viewer containing the digitised manuscript. For example: BL Add. 6613