BL: modifications to files previously published on FIHRIST (i.e. those on master branch)
Most of these modifications are to the files previously published on FIHRIST.

These are carried out on the bl-persian-mss-2022 branch to mimic parallel changes on the master branch (where the changes are described in more detail).

Most modifications are to tidy up the XML formatting and ensure that all files validate against two schemas. Also, a new standard availability statement is inserted.

Some duplicate comments were removed from the persons_base and works_base authority files.