BL: insert ref elements and locus text for miniatures in  21 manuscripts
The insertion of the <ref> element supports a hyperlink to the miniature in a digital surrogate.

The text of the locus provides the folio reference (prior to the summary description of the miniature).

(Prior to this change the folio reference was hidden and there was no direct link to the folio in the digitised ms)