Second attempt at building the FIHRIST dataset using master files from IAMS
We revert to the set of  2626 records on the master branch and then overwrite 2174 of those with 'master' records from IAMS.

In our first attempt at this we mistakenly used some master records that did not include manuscript_id in the TEI header (we used some IAMS records that have been independently catalogued in TEI by a Hebrew curator) e.g. Uk_Delhi_Persian_1003.