BL: 627 records modified after reconciliation of FIHRIST with IAMS
The changes are due to insertion of (IAMS) Arks alongside viaf_123 in @key values. These will be ignored by FIHRIST.

The Arks are inserted during IAMS (re)import but IAMS does not always do this in every occurrence. As a result, Arks can be inserted on subsequent re-imports to IAMS (as was the case here when availability @status values were recently modified).